requirements.txt
numpy
scipy
pandas
matplotlib
numba
tqdm
IPython
pillow
ipywidgets